Move the different Options to an own file for better maintainability
