+++ components/RarityScore.tsx
@@ -10,6 +10,8 @@ import {
     Typography
   } from '@iotabots/components'
 import { useEffect } from 'react';
+import {displayOptions, bodyOptions, faceOptions, backgroundOptions} from '../public/soonabotOptions.js'
+
 
 
 /**
@@ -54,30 +56,6 @@ import { useEffect } from 'react';
 //      )
 // }
 
-export const displayOptions  = [
-    { value: 0.03, name: 'Small' },
-    { value: 2.7, name: 'Medium' },
-    { value: 7.47, name: 'Large' },
-]
-
-export const bodyOptions  = [
-    { value: 0.41, name: 'Small' },
-    { value: 0.2, name: 'Medium' },
-    { value: 3.32, name: 'Large' },
-]
-
-export const faceOptions  = [
-    { value: 0.33, name: 'Small' },
-    { value: 2, name: 'Medium' },
-    { value: 5.19, name: 'Large' },
-]
-
-export const backgroundOptions  = [
-    { value: 5.09, name: 'Small' },
-    { value: 20, name: 'Medium' },
-    { value: 18.98, name: 'Large' },
-]
-
 
 export const RarityScoreSoonabots = () => {
     const [display, setDisplay] = useState(null);
@@ -85,7 +63,6 @@ export const RarityScoreSoonabots = () => {
     const [face, setFace] = useState(null);
     const [background, setBackground] = useState(null);
     const [result, setResult] = useState(null);
-
 
 
     const resetValues = () => {
