+++ public/soonabotOptions.js
@@ -1,0 +1,23 @@
+export const displayOptions  = [
+    { value: 0.03, name: 'Small' },
+    { value: 2.7, name: 'Medium' },
+    { value: 7.47, name: 'Large' },
+]
+
+export const bodyOptions  = [
+    { value: 0.41, name: 'Small' },
+    { value: 0.2, name: 'Medium' },
+    { value: 3.32, name: 'Large' },
+]
+
+export const faceOptions  = [
+    { value: 0.33, name: 'Small' },
+    { value: 2, name: 'Medium' },
+    { value: 5.19, name: 'Large' },
+]
+
+export const backgroundOptions  = [
+    { value: 5.09, name: 'Small' },
+    { value: 20, name: 'Medium' },
+    { value: 18.98, name: 'Large' },
+]
